# internal/openapi/openapi.yaml
openapi: 3.0.2
servers:
  - url: http://localhost:8080
    description: Local server
info:
  version: 1.0.0
  title: Users API
security:
  - ApiKeyAuth: []
tags:
  - name: Profiles
    description: User profile operations
paths:
  /v1/profiles:
    get:
      summary: Get all profiles
      description: Returns the list of user profiles
      operationId: GetProfiles
      tags:
        - Profiles
      security:
        - ApiKeyAuth: []
      parameters:
        - name: username
          in: query
          required: false
          schema:
            type: string
          description: Username of the user
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
          description: Number of profiles to return
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
          description: Number of profiles to skip
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProfilesResponse'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  schemas:
    Error:
      type: object
      required:
        - message
        - reason
      properties:
        message:
          type: string
          title: message
          description: Client facing error message
          example: Something went wrong
        reason:
          type: string
          title: reason
          description: Internal error
          example: Database connection error
    GetProfilesResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserProfile'
    UserProfile:
      required:
        - id
        - username
        - first_name
        - last_name
        - city
        - school
      properties:
        id:
          type: integer
          format: int32
          example: 7897432
          x-go-name: ID
        username:
          type: string
          example: johndoe
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        city:
          type: string
          example: New York
        school:
          type: string
          example: New York University